refactor(directory): revert back to initial directory pipeline
- move preprocess back to urlrepository
- add back rawdirectorysearch
- revert back tests
