+++ src/server/modules/user/services/__tests__/UrlManagementService.test.ts
@@ -21,8 +21,7 @@ describe('UrlManagementService', () => {
     findByShortUrl: jest.fn(),
     getLongUrl: jest.fn(),
     plainTextSearch: jest.fn(),
-    getRelevantUrlsFromEmail: jest.fn(),
-    getRelevantUrlsFromText: jest.fn(),
+    rawDirectorySearch: jest.fn(),
   }
 
   const service = new UrlManagementService(userRepository, urlRepository)
