+++ test/server/mocks/repositories/UrlRepository.ts
@@ -6,7 +6,11 @@ import {
   StorableFile,
   StorableUrl,
   UrlDirectoryPaginated,
+  UrlsPaginated,
 } from '../../../../src/server/repositories/types'
+import { StorableUrlState } from '../../../../src/server/repositories/enums'
+import { SearchResultsSortOrder } from '../../../../src/shared/search'
+import { DirectoryQueryConditions } from '../../../../src/server/services/interfaces/DirectorySearchServiceInterface'
 
 @injectable()
 export class UrlRepositoryMock implements UrlRepositoryInterface {
@@ -33,13 +37,8 @@ export class UrlRepositoryMock implements UrlRepositoryInterface {
     throw new Error('Not implemented')
   }
 
-  getRelevantUrlsFromEmail: (
-    likeQuery: string[],
-    rankingAlgorithm: string,
-    limit: number,
-    offset: number,
-    queryState: string[],
-    queryFile: boolean[],
+  rawDirectorySearch: (
+    condition: DirectoryQueryConditions,
   ) => Promise<UrlDirectoryPaginated> = () => {
     return Promise.resolve({
       urls: [
@@ -54,22 +53,24 @@ export class UrlRepositoryMock implements UrlRepositoryInterface {
     })
   }
 
-  getRelevantUrlsFromText: (
-    urlVector: string,
-    rankingAlgorithm: string,
+  plainTextSearch: (
+    query: string,
+    order: SearchResultsSortOrder,
     limit: number,
     offset: number,
-    query: string,
-    queryState: string,
-    queryFile: string,
-  ) => Promise<UrlDirectoryPaginated> = () => {
+  ) => Promise<UrlsPaginated> = () => {
     return Promise.resolve({
       urls: [
         {
           shortUrl: 'test-moh',
-          state: 'ACTIVE',
+          longUrl: 'https://www.moh.gov.sg/covid-19',
+          state: StorableUrlState.Active,
           isFile: false,
-          email: 'test@test.gov.sg',
+          createdAt: '2020-04-17T09:10:07.491Z',
+          updatedAt: '2020-06-09T10:07:07.557Z',
+          description: '',
+          contactEmail: null,
+          clicks: 0,
         },
       ],
       count: 0,
