+++ test/server/repositories/UrlRepository.test.ts
@@ -11,10 +11,11 @@ import {
 } from '../api/util'
 import { UrlRepository } from '../../../src/server/repositories/UrlRepository'
 import { UrlMapper } from '../../../src/server/mappers/UrlMapper'
+import { SearchResultsSortOrder } from '../../../src/shared/search'
 import { FileVisibility, S3ServerSide } from '../../../src/server/services/aws'
 import { NotFoundError } from '../../../src/server/util/error'
 import { StorableUrlState } from '../../../src/server/repositories/enums'
-import { urlSearchVector } from '../../../src/server/models/search'
+import { DirectoryQueryConditions } from '../../../src/server/services/interfaces/DirectorySearchServiceInterface'
 
 jest.mock('../../../src/server/models/url', () => ({
   Url: urlModelMock,
@@ -322,26 +323,22 @@ describe('UrlRepository', () => {
     })
   })
 
-  describe('directory search for email and keyword', () => {
+  describe('rawDirectorySearch', () => {
     beforeEach(() => {
       mockQuery.mockClear()
     })
-    it('should call sequelize.query that searches with email', async () => {
-      const likeQuery = ['@test.gov.sg', '@test2.gov.sg']
-      const rankingAlgorithm = `Url."createdAt"`
-      const limit = 100
-      const offset = 0
-      const queryState = ['ACTIVE']
-      const queryFile = [true]
+    it('should call sequelize.query that searches for emails', async () => {
+      const conditions: DirectoryQueryConditions = {
+        query: '@test.gov.sg @test2.gov.sg',
+        order: SearchResultsSortOrder.Recency,
+        limit: 100,
+        offset: 0,
+        state: 'ACTIVE',
+        isFile: true,
+        isEmail: true,
+      }
 
-      const repoawait = await repository.getRelevantUrlsFromEmail(
-        likeQuery,
-        rankingAlgorithm,
-        limit,
-        offset,
-        queryState,
-        queryFile,
-      )
+      const repoawait = await repository.rawDirectorySearch(conditions)
 
       expect(repoawait).toStrictEqual({
         count: 1,
@@ -357,23 +354,17 @@ describe('UrlRepository', () => {
     })
 
     it('should call sequelize.query that searches with plain text', async () => {
-      const urlVector = urlSearchVector
-      const query = 'query'
-      const rankingAlgorithm = `Url."createdAt"`
-      const limit = 100
-      const offset = 0
-      const queryState = 'ACTIVE'
-      const queryFile = ''
+      const conditions: DirectoryQueryConditions = {
+        query: 'query',
+        order: SearchResultsSortOrder.Recency,
+        limit: 100,
+        offset: 0,
+        state: 'ACTIVE',
+        isFile: true,
+        isEmail: false,
+      }
 
-      const repoawait = await repository.getRelevantUrlsFromText(
-        urlVector,
-        rankingAlgorithm,
-        limit,
-        offset,
-        query,
-        queryState,
-        queryFile,
-      )
+      const repoawait = await repository.rawDirectorySearch(conditions)
       expect(repoawait).toStrictEqual({
         count: 1,
         urls: [
