+++ test/server/services/DirectorySearchService.test.ts
@@ -2,36 +2,25 @@ import { DirectorySearchService } from '../../../src/server/services/DirectorySe
 import { UrlRepositoryMock } from '../mocks/repositories/UrlRepository'
 import { SearchResultsSortOrder } from '../../../src/shared/search'
 import { DirectoryQueryConditions } from '../../../src/server/services/interfaces/DirectorySearchServiceInterface'
-import { urlSearchVector } from '../../../src/server/models/search'
+
 /**
  * Unit tests for DirectorySearchService.
  */
-
-jest.mock('../../../src/server/models/url', () => ({
-  Url: { tableName: 'Url' },
-}))
-
-jest.mock('../../../src/server/util/parse', () => ({
-  sanitiseQuery: (query: string) => query,
-  extractShortUrl: (query: string) => query,
-}))
-
-const repository = new UrlRepositoryMock()
-const service = new DirectorySearchService(repository)
-
 describe('DirectorySearchService tests', () => {
   describe('plainTextSearch tests', () => {
-    it('Should return results from repository for keyword text search', async () => {
+    it('Should return results from repository', async () => {
+      const repository = new UrlRepositoryMock()
+      const service = new DirectorySearchService(repository)
       const conditions: DirectoryQueryConditions = {
         query: 'test-moh',
-        order: SearchResultsSortOrder.Recency,
+        order: SearchResultsSortOrder.Popularity,
         limit: 10,
         offset: 0,
         state: 'ACTIVE',
         isFile: false,
         isEmail: false,
       }
-      const spy = jest.spyOn(repository, 'getRelevantUrlsFromText')
+      const spy = jest.spyOn(repository, 'rawDirectorySearch')
       await expect(service.plainTextSearch(conditions)).resolves.toStrictEqual({
         urls: [
           {
@@ -43,50 +32,7 @@ describe('DirectorySearchService tests', () => {
         ],
         count: 0,
       })
-      expect(repository.getRelevantUrlsFromText).toBeCalledWith(
-        urlSearchVector,
-        // eslint-disable-next-line
-        "Url.\"createdAt\"",
-        10,
-        0,
-        // eslint-disable-next-line
-        "test-moh",
-        `AND urls.state = 'ACTIVE'`,
-        `AND urls."isFile"=false`,
-      )
-      spy.mockClear()
-    })
-    it('Should return results from repository for email search', async () => {
-      const conditions: DirectoryQueryConditions = {
-        query: '@test.gov.sg',
-        order: SearchResultsSortOrder.Recency,
-        limit: 10,
-        offset: 0,
-        state: 'ACTIVE',
-        isFile: false,
-        isEmail: true,
-      }
-      const spy = jest.spyOn(repository, 'getRelevantUrlsFromEmail')
-      await expect(service.plainTextSearch(conditions)).resolves.toStrictEqual({
-        urls: [
-          {
-            shortUrl: 'test-moh',
-            state: 'ACTIVE',
-            isFile: false,
-            email: 'test@test.gov.sg',
-          },
-        ],
-        count: 0,
-      })
-      expect(repository.getRelevantUrlsFromEmail).toBeCalledWith(
-        ['@test.gov.sg'],
-        // eslint-disable-next-line
-        "Url.\"createdAt\"",
-        10,
-        0,
-        ['ACTIVE'],
-        [false],
-      )
+      expect(repository.rawDirectorySearch).toBeCalledWith(conditions)
       spy.mockClear()
     })
   })
